+++ R/RunMAPOneSample.R
@@ -107,7 +107,8 @@ RunMAPOnOneSample <-
     reconstructed.spectrum <- retval$proposed.reconstruction
     colnames(reconstructed.spectrum) <-
       paste0("Reconstructed spectrum (count = ", round(colSums(reconstructed.spectrum)),
-             ", cosine similarity = ", round(distance.info$value["cosine"], 5), ")")
+             ", cosine similarity = ", 
+             round(distance.info$proposed.assignment["cosine"], 5), ")")
     colnames(spect) <- paste0(colnames(spect), " (count = ",colSums(spect), ")")
     list.of.catalogs <- list(spect, reconstructed.spectrum, sigs1)
     PlotListOfCatalogsToPdf(list.of.catalogs,